chore: centralise metrics fetching dates
